Gordon's changes

+++ examples/lighting-app/esp32/main/include/LEDCluster.h
@@ -41,7 +41,7 @@ public:
 
     void SetBrightness(uint8_t brightness);
     void UpdateState();
-    void SetColor(uint8_t Hue, uint8_t Saturation);
+    void SetColor(int16_t Hue, int16_t Saturation);
     void SetColorTemp(uint16_t temp);
     uint8_t GetLevel(void);
     bool IsTurnedOn(void);
